Make DynamicMailbox clonable
We swithced from holding the message sender in a boxed any to rolling our own
dynamic represention. This was done because we needed our Box<dyn Any + ...>
to also be clonable (since mailboxes should and are clonable by nature), but
that's not possible since you can't have an object-safe clonable any. Therefore
we opted to use raw pointers and TypeId ourselves to do the dynamic checking.

Dropping and cloning is done through careful management of the underlying Box
which is kept as a *mut u8 to elide type information, and is only ever
transmuted where the type is certain, either in the original constructor or
through a TypeId check.